backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
docker
python-dotenv
